feat: create a generic Data fetching hook
