+++ src/components/GameGrid.tsx
@@ -6,7 +6,7 @@ import GameCardContainer from "./GameCardContainer";
 
 
 const GameGrid = () => {
-  const { games, error, isLoading } = useGames()
+  const { data, error, isLoading } = useGames()
   const skeletons = [1, 2, 3, 4, 5, 6]
 
   return (
@@ -21,7 +21,7 @@ const GameGrid = () => {
           <GameCardContainer>
             <GameCardSkeleton key={skeleton} />
           </GameCardContainer>)}
-        {games.map(game =>
+        {data.map(game =>
           <GameCardContainer>
             <GameCard key={game.id} game={game} />
           </GameCardContainer>
