+++ src/hooks/useGenres.ts
@@ -1,40 +1,10 @@
-import { useEffect, useState } from 'react';
-import apiClient from '../services/api-client';
-import { CanceledError } from 'axios';
+import useData from './useData';
 
-interface Genre {
+export interface Genre {
   id: number;
   name: string;
 }
 
-interface fetchGenreResponse {
-  count: number;
-  results: Genre[];
-}
-
-const useGenres = () => {
-  const [genres, setGenres] = useState<Genre[]>([]);
-  const [error, setError] = useState('');
-  const [isLoading, setIsLoading] = useState(true);
-
-  useEffect(() => {
-    const controller = new AbortController();
-
-    apiClient<fetchGenreResponse>('/genres', { signal: controller.signal })
-      .then((res) => {
-        setGenres(res.data.results);
-        setIsLoading(false);
-      })
-      .catch((err) => {
-        if (err instanceof CanceledError) return;
-        setError(err.message);
-        setIsLoading(false);
-      });
-
-    return () => controller.abort();
-  }, []);
-
-  return { genres, error, isLoading };
-};
+const useGenres = () => useData<Genre>('/genres')
 
 export default useGenres;
